+++ client/tws_foundation_client/CHANGELOG.md
@@ -5,6 +5,9 @@
 - Notes: 
     1. Added Carriers service.
     1. Added vehicules_models service.
+    1. Changes on default date on datetime sets propeties.
+    1. Added validations on [Carrier] set clone method.
+    1. Added [_timestamp] initialization in default constructor method for [Plate] & [Truck] sets.
 
     1. [TrucksExternals] Create & update service methods added and tested.
     2. [Trucks] tests refactored.